refactor(api-gateway): split gql-rest helpers into focused modules
- move shared GraphQL REST logging into gql-rest-logger
- move constants into gql-rest-shared
- isolate multipart proxy logic in gql-rest-multipart to keep the main
route file concise
